+++ packages/infra/.env.sample
@@ -16,6 +16,5 @@ PORT=3000
 AWS_ACCOUNT_ID="1234567890"
 AWS_REGION="us-east-1"
 POSTMARK_API_KEY="xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
-ACM_CERTIFICATE_ARN="arn:aws:acm:region:account-id:certificate/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
 SESSION_PASSWORD_1="super-long-string-here-that-is-hard-to-guess-and-at-least-100-characters"
 MONGO_URL="mongodb+srv://USERNAME:PASSWORD@CLUSTER_NAME.CLUSTER_ID.mongodb.net/"
