--- packages/infra/utils/getAcmCertificate.ts
@@ -1,21 +1,0 @@
-import {
-  type ICertificate,
-  Certificate
-} from "aws-cdk-lib/aws-certificatemanager";
-import type { Stack } from "aws-cdk-lib";
-import { env } from "./env";
-
-type GetACMCertificateProps = {
-  stack: Stack;
-};
-
-export const getACMCertificate = ({
-  stack
-}: GetACMCertificateProps): ICertificate => {
-  const certificate = Certificate.fromCertificateArn(
-    stack,
-    "PlutomiCertificate",
-    env.ACM_CERTIFICATE_ARN
-  );
-  return certificate;
-};
