+++ packages/infra/utils/index.ts
@@ -4,4 +4,5 @@ export * from "./createTaskDefinition";
 export * from "./createTaskRole";
 export * from "./createVpc";
 export * from "./getHostedZone";
+export * from "./createCertificate";
 export * from "./env";
